requirements.txt
certifi==2019.9.11
chardet==3.0.4
ffmpeg-python==0.2.0
future==0.18.2
idna==2.8
pyvtt==0.0.2
requests==2.22.0
ujson==1.35
urllib3==1.25.7
youtube-dl==2019.11.5
